Fix implicit widening multiplication in Debug builds
The previous fixes only covered Release (#else) code paths. Debug
(#ifndef NDEBUG) blocks had unfixed `N * 1024` multiplications
assigned to size_t, triggering bugprone-implicit-widening-of-
multiplication-result. Also fix readability-container-size-empty
in merklecpp.h assert.

+++ test/compare_hash_functions.cpp
@@ -295,7 +295,7 @@ int main()
 #endif
 
 #ifndef NDEBUG
-    const size_t num_leaves = 128 * 1024;
+    const size_t num_leaves = static_cast<size_t>(128) * 1024;
     const size_t root_interval = 128;
 #else
     const size_t num_leaves = static_cast<size_t>(16) * 1024 * 1024;
